#29: refactors language toggle
